+++ test/private/createTempDir.test.mjs
@@ -1,12 +1,12 @@
 import { strictEqual } from 'assert';
+import { existsSync } from 'fs';
 import createTempDir from '../../private/createTempDir.mjs';
 import fsPathRemove from '../../private/fsPathRemove.mjs';
-import fsPathExists from '../fsPathExists.mjs';
 
 export default (tests) => {
   tests.add('`createTempDir`.', async () => {
     const tempDirPath = await createTempDir();
-    strictEqual(await fsPathExists(tempDirPath), true);
+    strictEqual(existsSync(tempDirPath), true);
     await fsPathRemove(tempDirPath);
   });
 };