+++ test/private/fsPathRemove.test.mjs
@@ -1,19 +1,19 @@
 import { rejects, strictEqual } from 'assert';
+import { existsSync } from 'fs';
 import createTempDir from '../../private/createTempDir.mjs';
 import fsPathRemove from '../../private/fsPathRemove.mjs';
-import fsPathExists from '../fsPathExists.mjs';
 
 export default (tests) => {
   tests.add('`fsPathRemove` with a directory that exists.', async () => {
     const tempDirPath = await createTempDir();
     await fsPathRemove(tempDirPath);
-    strictEqual(await fsPathExists(tempDirPath), false);
+    strictEqual(existsSync(tempDirPath), false);
   });
 
   tests.add('`fsPathRemove` with a directory that doesn’t exist.', async () => {
     const path = './this-directory-no-exists';
     await fsPathRemove(path);
-    strictEqual(await fsPathExists(path), false);
+    strictEqual(existsSync(path), false);
   });
 
   tests.add('`fsPathRemove` with an invalid path argument.', async () => {