+++ changelog.md
@@ -21,6 +21,7 @@
 - Simplified the ESLint config.
 - Removed `npm-debug.log` from the `.gitignore` file as npm [v4.2.0](https://github.com/npm/npm/releases/tag/v4.2.0)+ doesn’t create it in the current working directory.
 - Removed the test helper function `sleep`.
+- Replaced the test helper function `fsPathExists` with the Node.js `fs.existsSync` function.
 
 ## 3.0.0